+++ petpals/templates/_layout.html
@@ -69,7 +69,7 @@
           </div>
           <div class="d-flex align-items-center">
             <button type="button" class="btn btn-link text-dark px-3 me-2">
-              <a class="nav-item nav-link" href="{{ url_for('about') }}">About Us</a>
+              <a class="nav-item nav-link" href="{{ url_for('home_router.about') }}">About Us</a>
             </button>
             <!--If user is logged in only show logout, if not logged in show signup/login-->
             {% if current_user.is_authenticated %}
