+++ petpals/templates/forum.html
@@ -163,7 +163,7 @@
                     <div class="media-body">
                       <h6>
                         <a
-                          href="{{ url_for('post_router.post') }}"
+                          href="{{ url_for('forum_router.post_router.post') }}"
                           data-toggle="collapse"
                           data-target=".forum-content"
                           class="text-body"
@@ -208,7 +208,7 @@
                     <div class="media-body">
                       <h6>
                         <a
-                          href="{{ url_for('post_router.post') }}"
+                          href="{{ url_for('forum_router.post_router.post') }}"
                           data-toggle="collapse"
                           data-target=".forum-content"
                           class="text-body"
@@ -253,7 +253,7 @@
                     <div class="media-body">
                       <h6>
                         <a
-                          href="{{ url_for('post_router.post') }}"
+                          href="{{ url_for('forum_router.post_router.post') }}"
                           data-toggle="collapse"
                           data-target=".forum-content"
                           class="text-body"
@@ -298,7 +298,7 @@
                     <div class="media-body">
                       <h6>
                         <a
-                          href="{{ url_for('post_router.post') }}"
+                          href="{{ url_for('forum_router.post_router.post') }}"
                           data-toggle="collapse"
                           data-target=".forum-content"
                           class="text-body"
