+++ petpals/templates/index.html
@@ -18,7 +18,7 @@
       <h1 class="display-3 mt-3">Welcome to Pet Pals!</h1>
       <h2 class="m-4">Find a new Pal for your Pet!</h2>
       <p>
-        <a class="btn btn-info btn-lg" href="/faq" role="button"
+        <a class="btn btn-info btn-lg" href="{{ url_for('home_router.faq') }}" role="button"
           >Learn more &raquo;</a
         >
       </p>
@@ -71,7 +71,7 @@
 
       <div class="col-md-4">
         <p>
-          <a class="btn btn-secondary" href="/faq" role="button"
+          <a class="btn btn-secondary" href="{{ url_for('home_router.faq') }}" role="button"
             >View details &raquo;</a
           >
         </p>
