+++ petpals/templates/login.html
@@ -54,7 +54,7 @@
           </div>
 
           <small class="text-muted ml-2">
-            <a href="{{ url_for('reset_request') }}">Forgot Password?</a>
+            <a href="{{ url_for('reset_password_router.reset_request') }}">Forgot Password?</a>
           </small>
         </form>
       </div>
